Atomic chat skills: split monolithic chat-with-agent into chat-sdk, detect-mode, chat-directline
Replace the monolithic chat-with-agent.js script (which hardcoded auth mode → protocol
mapping internally) with three atomic scripts where the LLM decides which to invoke:

- chat-sdk.bundle.js: SDK/M365 path only, works with any auth mode
- detect-mode.bundle.js: queries Dataverse for auth mode, returns recommendation
- directline-chat.bundle.js: already existed, now used by chat-directline skill

This enables using the M365 Agents SDK with no-auth agents (e.g., S2S D2E via SPN),
which was previously impossible because the monolithic script forced DirectLine for
authMode=None.

Removed:
- scripts/src/chat-with-agent.js and its bundle
- skills/chat-with-agent (deprecated, pointed to monolithic script)
- skills/directline-chat (deprecated, replaced by chat-directline)

Updated:
- skills/chat-sdk: points to chat-sdk.bundle.js, broadened description
- skills/chat-directline: points to directline-chat.bundle.js
- skills/detect-mode: points to detect-mode.bundle.js, returns recommendedMode
- agents/copilot-studio-test.md: allows user override of protocol choice
