+++ scripts/package.json
@@ -4,9 +4,10 @@
   "private": true,
   "description": "Bundled scripts for Copilot Studio plugin",
   "scripts": {
-    "build": "npm run build:schema-lookup && npm run build:chat-with-agent && npm run build:connector-lookup && npm run build:directline-chat && npm run build:manage-agent",
+    "build": "npm run build:schema-lookup && npm run build:chat-sdk && npm run build:detect-mode && npm run build:connector-lookup && npm run build:directline-chat && npm run build:manage-agent",
     "build:schema-lookup": "esbuild src/schema-lookup.js --bundle --platform=node --target=node18 --outfile=schema-lookup.bundle.js --format=cjs",
-    "build:chat-with-agent": "esbuild src/chat-with-agent.js --bundle --platform=node --target=node18 --outfile=chat-with-agent.bundle.js --format=cjs --external:keytar --external:@azure/msal-node-extensions --banner:js=\"var _d=process.env.CLAUDE_PLUGIN_DATA;if(!_d){try{_d=JSON.parse(require('fs').readFileSync(require('path').join(require('os').homedir(),'.copilot-studio-cli','plugin-paths.json'),'utf8')).pluginData}catch{}}if(_d){var _p=require('path');process.env.NODE_PATH=[_p.join(_d,'node_modules'),process.env.NODE_PATH].filter(Boolean).join(_p.delimiter);require('module')._initPaths()}\"",
+    "build:chat-sdk": "esbuild src/chat-sdk.js --bundle --platform=node --target=node18 --outfile=chat-sdk.bundle.js --format=cjs --external:keytar --external:@azure/msal-node-extensions --banner:js=\"var _d=process.env.CLAUDE_PLUGIN_DATA;if(!_d){try{_d=JSON.parse(require('fs').readFileSync(require('path').join(require('os').homedir(),'.copilot-studio-cli','plugin-paths.json'),'utf8')).pluginData}catch{}}if(_d){var _p=require('path');process.env.NODE_PATH=[_p.join(_d,'node_modules'),process.env.NODE_PATH].filter(Boolean).join(_p.delimiter);require('module')._initPaths()}\"",
+    "build:detect-mode": "esbuild src/detect-mode.js --bundle --platform=node --target=node18 --outfile=detect-mode.bundle.js --format=cjs --external:keytar --external:@azure/msal-node-extensions --banner:js=\"var _d=process.env.CLAUDE_PLUGIN_DATA;if(!_d){try{_d=JSON.parse(require('fs').readFileSync(require('path').join(require('os').homedir(),'.copilot-studio-cli','plugin-paths.json'),'utf8')).pluginData}catch{}}if(_d){var _p=require('path');process.env.NODE_PATH=[_p.join(_d,'node_modules'),process.env.NODE_PATH].filter(Boolean).join(_p.delimiter);require('module')._initPaths()}\"",
     "build:connector-lookup": "esbuild src/connector-lookup.js --bundle --platform=node --target=node18 --outfile=connector-lookup.bundle.js --format=cjs",
     "build:directline-chat": "esbuild src/directline-chat.js --bundle --platform=node --target=node18 --outfile=directline-chat.bundle.js --format=cjs",
     "build:manage-agent": "esbuild src/manage-agent.js --bundle --platform=node --target=node18 --outfile=manage-agent.bundle.js --format=cjs --external:keytar --external:@azure/msal-node-extensions --external:vscode-jsonrpc --banner:js=\"var _d=process.env.CLAUDE_PLUGIN_DATA;if(!_d){try{_d=JSON.parse(require('fs').readFileSync(require('path').join(require('os').homedir(),'.copilot-studio-cli','plugin-paths.json'),'utf8')).pluginData}catch{}}if(_d){var _p=require('path');process.env.NODE_PATH=[_p.join(_d,'node_modules'),process.env.NODE_PATH].filter(Boolean).join(_p.delimiter);require('module')._initPaths()}\"",
