+++ skills/chat-directline/SKILL.md
@@ -2,27 +2,27 @@
 user-invocable: false
 description: Send a message to a Copilot Studio agent via DirectLine v3. Use for agents with no auth or manual auth. Requires a token endpoint URL or DirectLine secret.
 argument-hint: <utterance>
-allowed-tools: Bash(node *chat-with-agent.bundle.js *), Read, Glob
+allowed-tools: Bash(node *directline-chat.bundle.js *), Read, Glob
 agent: copilot-studio-test
 ---
 
 # Chat via DirectLine
 
-Send a single utterance to a published agent via the DirectLine v3 REST API. Use this for agents with **no authentication** (authenticationmode 1) or **manual authentication** (authenticationmode 3).
+Send a single utterance to a published agent via the DirectLine v3 REST API. Typically used for agents with **no authentication** (authenticationmode 1) or **manual authentication** (authenticationmode 3).
 
 ## Instructions
 
 1. Run the script with `--token-endpoint` (from detect-mode output) and the utterance from `$ARGUMENTS`:
 
    ```bash
-   node ${CLAUDE_SKILL_DIR}/../../scripts/chat-with-agent.bundle.js \
+   node ${CLAUDE_SKILL_DIR}/../../scripts/directline-chat.bundle.js \
      --token-endpoint "<tokenEndpoint>" "<utterance>"
    ```
 
    If the user provided a DirectLine secret instead:
 
    ```bash
-   node ${CLAUDE_SKILL_DIR}/../../scripts/chat-with-agent.bundle.js \
+   node ${CLAUDE_SKILL_DIR}/../../scripts/directline-chat.bundle.js \
      --directline-secret "<secret>" "<utterance>"
    ```
 
@@ -65,8 +65,8 @@ The agent requires the user to authenticate (manual auth agents):
 ```
 
 1. Show the `signin_url` to the user, ask for the validation code
-2. Send the code: `node ${CLAUDE_SKILL_DIR}/../../scripts/chat-with-agent.bundle.js <resume_command with code substituted>`
-3. Re-send the utterance: `node ${CLAUDE_SKILL_DIR}/../../scripts/chat-with-agent.bundle.js <followup_command>`
+2. Send the code: `node ${CLAUDE_SKILL_DIR}/../../scripts/directline-chat.bundle.js <resume_command with code substituted>`
+3. Re-send the utterance: `node ${CLAUDE_SKILL_DIR}/../../scripts/directline-chat.bundle.js <followup_command>`
 
 Use `resume_command` and `followup_command` exactly as given.
 
@@ -75,7 +75,7 @@ Use `resume_command` and `followup_command` exactly as given.
 Pass all three values from the previous response:
 
 ```bash
-node ${CLAUDE_SKILL_DIR}/../../scripts/chat-with-agent.bundle.js \
+node ${CLAUDE_SKILL_DIR}/../../scripts/directline-chat.bundle.js \
   --token-endpoint "<url>" "<follow-up>" \
   --conversation-id "<id>" --directline-token "<token>" --watermark "<watermark>"
 ```
